switch to 2nd variant
